backend/requirements.txt
aniso8601==9.0.1
apispec==6.0.2
apispec-webframeworks==0.5.2
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
Flask-MySQLdb==1.0.1
flask-swagger-ui==4.11.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.29
mysqlclient==2.1.1
packaging==22.0
protobuf==3.20.1
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.7
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.14
validate-email==1.3
Werkzeug==2.2.2
